COZMO-13139 Use size_t for Matrix dimensions
The previous code used s32 for Matrix dimensions which caused template
deduction problems in Android NDK r15b's version of clang.
Specifically, clang was treating naked numbers as unsigned ints instead
of being flexible and considering them as signed ints.  As a result,
various function templates would not be matched.

+++ coretech/common/include/anki/common/basestation/math/matrix_impl.h
@@ -356,7 +356,7 @@ namespace Anki {
   {
     CORETECH_THROW_IF(i >= NROWS || j >= NCOLS);
 #if ANKICORETECH_USE_OPENCV
-    return cv::Matx<T,NROWS,NCOLS>::operator()(i,j);
+    return cv::Matx<T,NROWS,NCOLS>::operator()((int) i, (int) j);
 #else
     CORETECH_ASSERT(false);
     // TODO: Define our own opencv-free (i,j) accessor?
@@ -370,7 +370,7 @@ namespace Anki {
   {
     CORETECH_THROW_IF(i >= NROWS || j >= NCOLS);
 #if ANKICORETECH_USE_OPENCV
-    return cv::Matx<T,NROWS,NCOLS>::operator()(i,j);
+    return cv::Matx<T,NROWS,NCOLS>::operator()((int) i, (int)j);
 #else
     CORETECH_ASSERT(false);
     // TODO: Define our own opencv-free const (i,j) accessor?
